+++ sample/src/main/java/com/airwallex/paymentacceptance/repo/BaseRepository.kt
@@ -4,10 +4,10 @@ import com.airwallex.android.core.model.PaymentIntent
 
 interface BaseRepository {
 
-    suspend fun login()
+    suspend fun login() = Unit
 
     suspend fun getPaymentIntentFromServer(
-        force3DS: Boolean = false,
+        force3DS: Boolean? = false,
         customerId: String? = null
     ): PaymentIntent
