+++ sample/src/main/java/com/airwallex/paymentacceptance/repo/RepositoryProvider.kt
@@ -1,0 +1,8 @@
+package com.airwallex.paymentacceptance.repo
+
+object RepositoryProvider {
+
+    fun get(): BaseRepository {
+        return PACheckoutDemoRepository()
+    }
+}
